+++ Index.php
@@ -4,5 +4,4 @@ declare(strict_types=1);
 
 namespace Usi;
 
-require_once(sprintf("%s/Infrastructure/Session.php", $_SERVER["DOCUMENT_ROOT"]));
 require_once(sprintf("%s/Home/index.php", $_SERVER["DOCUMENT_ROOT"]));
