+++ assets/configuration/DEV/environment.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<environment xmlns="http://usi.gov.au/ws" url="https://dev.portal.usi.gov.au/service/v5/usiservice.svc" defaultOrgCode="VA1803">
+<environment xmlns="http://usi.gov.au/ws" url="https://dev.portal.usi.gov.au/service/v5/usiservice.svc" defaultOrgCode="0002">
     <sts url="https://softwareauthorisations.acc.ato.gov.au/R3.0/S007v1.3/service.svc" uri="https://3pt.portal.usi.gov.au/service/usiservice.svc">
         <keyStoreMapping>
             <add abn="27809366375" code="0002" privateKeyPassword="Password1!" />
