+++ assets/js/home.js
@@ -1,5 +1,5 @@
 $(document).ready(function () {
-    $(".content .language-xml").each(function () {
+    $(".language-xml").each(function () {
         hljs.highlightElement(this);
     });
 });
